Don't reset CMAKE_BUILD_TYPE if already set
yubihsm's cmake file will set CMAKE_BUILD_TYPE based on the value of RELEASE_BUILD (presumably typically set via -DRELEASE_BUILD=1 on the cmake command line). This is a bit unorthodox and causes problems in two situations: 1) a user familiar with cmake will -DCMAKE_BUILD_TYPE=Release only to have it silently set back to Debug from under him or her, and 2) when integrating the library in to a larger cmake tree the yubihsm directories will ignore the "global" setting.

Change the cmake file to not reset CMAKE_BUILD_TYPE if it's already set.