Add a library only build option
When integrating the yubihsm-shell repository in to other cmake projects oftentimes only the yubihsm library is required. However, the CMakeLists.txt will bring in a number of dependencies needed for the various tools, such as libedit, help2man, etc., that are unneeded for just building the library.

Add a cmake option to enable a library only build so that these additional dependencies are not required by cmake when a parent cmake project just wants to link against libyubihsm.