tool: stop the timer for keepalive while we do EVP_read_pw_string()
openssl gets sad if a signal hits while that function is waiting for
input from the user and will both be interrupted and the actual input
will be dumped on the prompt.

Also move the blocking of keepalive further into when we actually make a
call to the device rather than being around the whole collection and
validation logic.