cmake: Fix build against pcsc-lite >= 2.2
The pcsc-lite 2.2.0 switched from autotools to meson and reworked the
pkgconfig files. The new pkg config provides CFLAGS that work ok, but
the yubihsm-shell ignores them and hopes that all included files are in
the default include directory (with the PCSC prefix).

Note, the value ${LIBPCSC_CFLAGS} is a semicolon separated list which
we need to split to separate items here.

This solution works with both old and new versions.

Fixes: #404

Signed-off-by: Jakub Jelen <jjelen@redhat.com>