tests: Fix untermined string initializaions
The new GCC 15 reports error when the string initializers overflow the
size of the underlying structure. This is common when the byte strings
are constructed in quotes such as "\xBB" as such this string has
trailing null byte and therefore the size two.

This is not an issue in the tests as they do not expect the string to be
NULL terminated, but it might uncover issues in other cases.

Example of the error:

/builddir/build/BUILD/yubihsm-shell-2.6.0-build/yubihsm-shell-2.6.0/pkcs11/tests/aes_encrypt_test.c:38:3: error: initializer-string for array of ‘unsigned char’ is too long [-Werror=unterminated-string-initialization]
   38 |   "\x6b\xc1\xbe\xe2\x2e\x40\x9f\x96\xe9\x3d\x7e\x11\x73\x93\x17\x2a"
      |   ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

Signed-off-by: Jakub Jelen <jjelen@redhat.com>